requirements.txt
colorama==0.4.4
crytic_compile>=0.3.1,<0.4.0
eth_typing>=3.0.0
eth_utils>=2.1.0
slither_analyzer==0.9.3
web3==6.0.0
requests~=2.28.2
setuptools~=65.5.1